update code - using doubleValue method
- 반환된 타입이 BigDecimal 객체를 doubleValue() 메서드를 사용하여 double 값으로 변환한다.
- 추후 다른 메서드 타입도 BigDecimal 타입으로 변경하기 위해서 임시적으로 doubleValue()를 호출함

+++ src/main/java/domain/amount/MonthlyInvestmentAmount.java
@@ -23,7 +23,7 @@ public class MonthlyInvestmentAmount implements TargetAmountReachable {
 
 	@Override
 	public int calMonthsToReach(int initialCapital, TargetAmount targetAmount, InterestRate interestRate) {
-		double monthlyRate = interestRate.getMonthlyRate();
+		double monthlyRate = interestRate.getMonthlyRate().doubleValue();
 		int months = 0;
 		double balance = initialCapital;
 		while (balance < targetAmount.getAmount()) {
